Time Series Forecasting with Lag-Llama
In this tutorial, you'll use the Lag-Llama model and see how it does in two different forecasting tasks. Zero-shot learning is when the model is not trained on the data it’s trying to predict. It’s an interesting test of how well the model can detect and respond to the patterns present in the time series. After that, you’ll fine-tune the model to see whether there’s a performance boost. On the free tier of Watsonx fine-tuning might take up to an hour but the model can be saved after that step for use later. The paid tier of watsonx provides access to a GPU that will decrease the fine-tuning time to around 10 minutes. With that in mind, let's get started.